@@ -85,7 +85,11 @@ AppContext.SetSwitch("Npgsql.EnableLegacyTimestampBehavior", true);
 
 using var scope = app.Services.CreateScope();
 var dbContext = scope.ServiceProvider.GetRequiredService<AuthenticationDbContext>();
-await dbContext.Database.MigrateAsync().ConfigureAwait(false);
+var migrations = await dbContext.Database.GetPendingMigrationsAsync();
+if (migrations.Any())
+{
+    await dbContext.Database.MigrateAsync().ConfigureAwait(false);
+}
 
 
 // Configure the HTTP request pipeline.
